// frontend-admin/src/api/generated/model/tenantDeleteDependencyCountsDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface TenantDeleteDependencyCountsDto {
  auditLogs?: number;
  cashRegisters?: number;
  categories?: number;
  dailyClosings?: number;
  finanzOnlineSubmissions?: number;
  memberships?: number;
  payments?: number;
  products?: number;
  receipts?: number;
  users?: number;
  voucherLedgerEntries?: number;
  vouchers?: number;
}